Makes the usage of EventBuffer way easier.
